refactor(permission): 디렉토리 분할
