+++ src/main/java/com/simpleboard/board/permission/infrastructure/entity/ManagerAssignmentEntity.java
@@ -1,7 +1,7 @@
 package com.simpleboard.board.permission.infrastructure.entity;
 
-import com.simpleboard.board.permission.domain.ManagerAssignment;
-import com.simpleboard.board.permission.domain.RoleName;
+import com.simpleboard.board.permission.domain.entity.ManagerAssignment;
+import com.simpleboard.board.permission.domain.vo.RoleName;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
