+++ src/test/java/com/simpleboard/board/permission/infrastructure/entity/PermissionPolicyEntityTest.java
@@ -2,9 +2,9 @@ package com.simpleboard.board.permission.infrastructure.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.simpleboard.board.permission.domain.ManagerAssignment;
-import com.simpleboard.board.permission.domain.PermissionPolicy;
-import com.simpleboard.board.permission.domain.RoleName;
+import com.simpleboard.board.permission.domain.entity.ManagerAssignment;
+import com.simpleboard.board.permission.domain.entity.PermissionPolicy;
+import com.simpleboard.board.permission.domain.vo.RoleName;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
